requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
numpy==2.4.4
openpyxl==3.1.5
pandas==3.0.2
pillow==12.2.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
